+++ TaleEngine/TaleEngine.IntegrationTests/Base/TaleEngineScenarioBase.cs
@@ -1,0 +1,25 @@
+﻿using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.TestHost;
+using System.Diagnostics.CodeAnalysis;
+using Microsoft.Extensions.Configuration;
+
+namespace TaleEngine.IntegrationTests.Base
+{
+    [ExcludeFromCodeCoverage]
+    public class TaleEngineScenarioBase
+    {
+        public TestServer CreateServer()
+        {
+            var hostBuilder = new WebHostBuilder()
+                .ConfigureAppConfiguration(cb =>
+                {
+                    cb.AddJsonFile("appsettings.json", optional: false)
+                    .AddEnvironmentVariables();
+                })
+                .UseEnvironment("IntegrationTest")
+                .UseStartup<Startup>();
+
+            return new TestServer(hostBuilder);
+        }
+    }
+}
