+++ TaleEngine/TaleEngine.IntegrationTests/appsettings.json
@@ -5,7 +5,5 @@
     }
   },
   "AllowedHosts": "*",
-  "ConnectionStrings": {
-    "DefaultConnection": "Server=localhost;Database=TaleEngine;Trusted_Connection=True;"
-  }
+  "ConnectionString": "Server=localhost;Database=TaleEngine;Trusted_Connection=True;"
 }
